refactor: 빌드 로직 개선 및 `build-logic` 구조 변경
- `app` 모듈의 `buildTypes` 설정을 `build-logic`의 `AndroidApplicationConventionPlugin`으로 이전하여 중복을 제거하고 설정을 중앙화했습니다.
- `build-logic` 내부에 `internal` 패키지를 추가하고, 기존의 최상위 파일들(`ApplicationConfig`, `ConfigureAndroid`, `ConfigureCompose`, `Extension`, `Plugins`)을 이 패키지 안으로 이동시켜 구조를 개선했습니다.
- 각 모듈의 `build.gradle.kts` 파일에서 불필요하게 선언되었던 의존성 및 ProGuard 설정을 제거하고, Convention Plugin을 통해 관리하도록 수정했습니다.
